Updated setup script to not overwrite Xresources aliases

+++ setup.sh
@@ -2,8 +2,13 @@
 
 # =1= .Xresources alias files.
 # Setup .Xdefaults symlink to .Xresources
-ln -s .Xresources .Xdefaults
+if [ ! -f ".Xdefaults" ]; then
+	ln -s .Xresources .Xdefaults
+fi
 #Setup .Xdefaults-$host symlink to .Xdefaults (xterm has a quirk :) )...
-ln -s .Xdefaults ".Xdefaults-$(uname -n)"
+xdef_xterm_alias=".Xdefaults-$(uname -n)"
+if [ ! -f "$xdef_xterm_alias" ]; then
+	ln -s .Xdefaults $xdef_xterm_alias
+fi
